One ffmpeg pass, however many Definition Cards
compositeOverlaysAtOffsets takes the video plus N rendered .mov files and
encodes ONCE: the graph shifts each card onto its own moment and gates it
there, so a second Definition Card adds a filter node, never another pass.

Deliberately a new method rather than a widening of compositeOverlay: that one
is the vertical Shorts pipeline's single, full-length subtitle track, and its
bytes must not move. They share the encode settings — including the bitexact
flags every output pass owes the Byte Hash — and nothing else.

ffmpeg-commands.ts was already over the file-token limit before this branch, so
nothing could be added to it at all. runFfmpegWithProgress, the bitexact args
and FFmpegError move out to ffmpeg-run.ts unchanged: they are the invariant
every ffmpeg call in this app owes, and the commands are the variations on it.
